AI Pathfinding implemented!
Yay! It works! NOTE: There's a bug that occurs sometimes and I don't know why. Sometimes the HERO will finish walking, only to jump back to the previous cell, and all the co-ords get messed up as far as I can tell. Making the character move after that will result in a crash.