Redesigned unit stats, fixed unit selection
The unit stats system was redesigned to be more friendly and make use of default stats. Now when creating a unit we need only give it the stats we care about, and everything else will remain the default. We no longer need to have an argument for each stat type when creating a new unit. Enemies can share this stat system, which should in turn lead to less repetitive code and more stability (potentially).

I also fixed the targetHero still being selected after moving.