New pathfinding system, AI Behaviour, and camera system
The old pathfinding system didn't allow for the NPCs/Enemies to only move a certain distance based on their movement range. This required a lot of reworking, and we may still replace it with a fully customized A* later if needed.

A preliminary AI behavioural system has been added too, seems to work well. It's somewhat similar to the Gambit system in Final Fantasy 12.

The camera is now smoother and can be told who/what to follow and it keeps them in the frame properly. I may further tweak it to not let it go to far outside the map.