Player Heroes can now attack enemies (mostly)
Still need to end the action and make it so we can't do anything else with that unit as they've performed an action.

The endTurn() method seems to affect the CURSOR instead of all the units. This could be fine if we're going use an initiative based system rather than allowing players to move all their units on their turn, but since we're undecided on that we need to change that method to apply to the selected unit.