Fixed depth sorting
Problems with depth sorting turned out to be a result of a temporary creation process for building random walls, where I created random walls between grid cell 10 and width/height -10, but used random() instead of irandom() and so was getting floating points numbers for the x and y coords. They appeared normal because the drawing performs its own calculation based on the grid position, not x and y.