Started implementing fire
Fire can be spawned and occupies its own grid layer called FIRE_GRID.

Fire instances have health and each turn spawn more fire objects next to themselves with 1 less health, so the fire eventually dies out.

Oddly, the fire doesn't all disappear in one go as I expected it to, some sticks around inexplicably. It's not a bad thing and could be implemented deliberately as a feature, but right now it's just a bug.

Fire are objects and technically NPCs and have turns, so drawing the turn list was modified to incorporate a new variable showOnTurnList. This works fine but involves copying the turnlist and filtering it every step! This can easily be optimized by making it only happen when a change occurs in the turns.

To do:
Fire needs to consider the terrain it's going to spread to and also do it damage each turn. This means deciding on how we're going to approach the floor; as objects or just a grid with values. Objects allow for more flexibility and easier implementation, but a grid system keeps things working fast, so that's probably the thing to try first, but it needs brainstorming.