Fixed an issue where fire would be seen as a valid target
The enemyPirate was walking towards the fire as it deemed it a valid target. Fixed this by putting FIRE on its own team and excluding that team as potential enemies. This may not be ideal in the long run as we'll probably want mages who can extinguish fires, but we can cross that bridge when we get to it.