AI will now avoid fire and walk away from it
Right now they only move to the nearest safe panel, we might want to change this to the safest panel, but that would require a lot more calculations and code. For now this is fine.

For some reason now the AI is only moving a maximum of 1 panel at a time. I think this is due to the code I commented out in the ENEMY_AI movement scripting in the step event (this was done because there was not listOfCells list when using AI_EscapeFire(). We probably need to refactor how we're going about movement there, might need to make an exception, but that could get messy as this is probably going to come up a lot.