Started rewriting the weapon system
I didn't like using tables to store weapon data, especially default stats. Tables will still likely be used for storing the persistent weapons, but for randomly generated weapons I liked just using objects themselves. There's now a simple system that will generate a random SWORD which in turn generates a FOIL, which in turn generates an Epee or Rapier, or you can be more specific and just generate a random Rapier if that's what you want. It could be modified to generate a random WEAPON which I might do for loot generation (which I suppose could inevitably be LOOT, but the tradeoff is that we need to put all the possible children objects into a list, which could become hard to maintain).