Continued development of custom pathfinding
Using NODE objects would have created too much overhead so I changed the system to use hash maps as pseudo-objects instead. This will be much faster and much less memory impactful (assuming we garbage collect properly) but integration is proving a bit trickier than it was to make in a dedicated test project.