Updating for making public.
The current state has issues with the camera being jittery. Things used to be tied to delta time so I could try to allow infinite frame rates, but that ended up being undesirable for a turn based game. I need to investigate this issue more when I come back to this project.